Recover TGeoManager::fIsOutside for transf. tracks
When changing from one engine to the other, the TGeoManager's state
flagging if the geometry has been left might be still set to
fIsOutside == true.
That can be a remnant from the transportation of the former engine. This
state is now cached when tracks are transferred and recovered when
tracks are picked up for tracking again.

TVirtualMC can request the recovery of a geometry state via calling
TVirtualMC::TryRestoreGeometryState(Int_t trackId).
Every further touch of the TGeoManager/TGeoNavigator within the
requesting engine is done in its responsibility.