Finalise geometry recovery
The recovery of the previous geometry state of a transferred track can
be requested from the TMCManager.
A VMC engine can hence
either call
TMCManager::RestoreGeometryState(Int_t trackId, ...) to request the
recovery for a specified track by its ID
or
TMCManager::RestoreGeometryState() to request the recovery of the track
which is currently set on the current VMC stack.