Recover TGeoManager::fIsOutside for transfer tracks (#3)
- Added functions to TMCManager:
    Bool_t RestoreGeometryState(Int_t trackId, Bool_t checkTrackIdRange = kTRUE);
    Bool_t RestoreGeometryState();
- Added TMCParticleStatus::fIsOutside
- Removed function from TMCManagerStack (not used)
   void NotifyOnRestoredGeometry(Int_t trackId);
   void NotifyOnRestoredGeometry();

From Benedikt:
When changing from one engine to the other, the TGeoManager's state flagging if the geometry has been left might be still set to 
   fIsOutside == true.
That can be a remnant from the transportation of the former engine. This state is now cached  when tracks are transferred and recovered when tracks are picked up for tracking again.

The recovery of the previous geometry state of a transferred track can be requested from the TMCManager. A VMC engine can hence either call
   TMCManager::RestoreGeometryState(Int_t trackId, ...) 
to request the recovery for a specified track by its ID or
   TMCManager::RestoreGeometryState() to request the recovery of the track
which is currently set on the current VMC stack.